requirements.txt
astunparse==1.6.3
av==10.0.0
backports.statistics==0.1.0
beartype==0.14.0
brotli==1.0.9
brotlicffi==1.0.9.2
chardet==5.1.0
colorama==0.4.6
conda==4.3.16
coremltools==6.3.0
cryptography==41.0.1
Cython==0.29.35
defusedxml==0.7.1
dill==0.3.6
dl==0.1.0
dp==1.00
einops==0.6.1
etcd==2.0.8
expecttest==0.1.4
fast_bss_eval==0.1.4
flashlight==0.1.1
foo==.1
genapi==0.0.8
h5py==3.8.0
hypothesis==6.76.0
importlib_metadata==6.6.0
ipaddr==2.2.0
ipython==8.14.0
keyring==23.13.1
lmdb==1.4.1
lpips==0.1.4
matplotlib==3.7.1
mock==5.0.2
monkeytype==23.3.0
moviepy==1.0.3
mtrand==0.1
mypy==1.3.0
mypy_extensions==1.0.0
networkx==3.1
nltk==3.8.1
nose==1.3.7
numarray==1.5.1
Numeric==24.2
olefile==0.46
onnxruntime==1.15.0
ont_fast5_api==4.1.1
optuna==3.2.0
ordereddict==1.1
pandas==2.0.2
pickle5==0.0.12
protobuf==4.23.2
psutil==5.9.5
pycocotools==2.0.6
pydot==1.4.2
PyInstaller==5.11.0
pynvml==11.5.0
pyOpenSSL==23.2.0
PyQt5==5.15.9
PyQt6==6.5.1
PySide2==5.15.2.1
PySide6==6.5.1
pystoi==0.3.3
pytest==7.3.1
pytorch_lightning==2.0.2
pytz==2023.3
PyYAML==6.0
ray==2.4.0
regex==2023.6.3
scikit_learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentencepiece==0.1.99
simplejson==3.19.1
soundfile==0.12.1
sympy==1.12
tabulate==0.9.0
tensorflow==2.12.0
threadpoolctl==3.1.0
torch_fidelity==0.3.0
torcharrow==0.1.0
torchdynamo==1.13.0
tqdm==4.65.0
transformers==4.29.2
urllib3_secure_extra==0.1.0
wandb==0.15.3
wincertstore==0.2.1
xmlrunner==1.7.7
z3==0.2.0
zstandard==0.21.0
